isolate deep link functions in appropriate file
