+++ vscode/microsoft-kiota/src/steps.ts
@@ -2,8 +2,8 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { Disposable, l10n, OpenDialogOptions, QuickInput, QuickInputButton, QuickInputButtons, QuickPickItem, Uri, window, workspace } from 'vscode';
 import { allGenerationLanguages, generationLanguageToString, KiotaSearchResultItem, LanguagesInformation, maturityLevelToString } from './kiotaInterop';
-import { findAppPackageDirectory, getWorkspaceJsonDirectory, IntegrationParams } from './util';
-import { createTemporaryFolder, isTemporaryDirectory } from './utilities/temporary-folder';
+import { findAppPackageDirectory, getWorkspaceJsonDirectory } from './util';
+import { isTemporaryDirectory } from './utilities/temporary-folder';
 
 export async function filterSteps(existingFilter: string, filterCallback: (searchQuery: string) => void) {
     const state = {} as Partial<BaseStepsState>;
@@ -116,38 +116,6 @@ export async function searchSteps(searchCallBack: (searchQuery: string) => Thena
     await MultiStepInput.run(input => inputPathOrSearch(input, state), () => step -= 2);
     return state;
 }
-
-export function transformToGenerationconfig(deepLinkParams: Partial<IntegrationParams>)
-    : Partial<GenerateState> {
-    const generationConfig: Partial<GenerateState> = {};
-    if (deepLinkParams.kind === "client") {
-        generationConfig.generationType = "client";
-        generationConfig.clientClassName = deepLinkParams.name;
-        generationConfig.language = deepLinkParams.language;
-    }
-    else if (deepLinkParams.kind === "plugin") {
-        generationConfig.pluginName = deepLinkParams.name;
-        switch (deepLinkParams.type) {
-            case "apiplugin":
-                generationConfig.generationType = "plugin";
-                generationConfig.pluginTypes = ["ApiPlugin"];
-                break;
-            case "openai":
-                generationConfig.generationType = "plugin";
-                generationConfig.pluginTypes = ['OpenAI'];
-                break;
-            case "apimanifest":
-                generationConfig.generationType = "apimanifest";
-                break;
-        }
-        generationConfig.outputPath =
-            (deepLinkParams.source && deepLinkParams.source?.toLowerCase() === 'ttk')
-                ? createTemporaryFolder()
-                : undefined;
-    }
-    return generationConfig;
-}
-
 
 export async function generateSteps(existingConfiguration: Partial<GenerateState>, languagesInformation?: LanguagesInformation, isDeeplinkEnabled?: boolean) {
     const state = { ...existingConfiguration } as Partial<GenerateState>;
@@ -389,7 +357,7 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
         }
     }
     async function inputManifestName(input: MultiStepInput, state: Partial<GenerateState>) {
-        if (!isDeepLinkPluginNameProvided)  {
+        if (!isDeepLinkPluginNameProvided) {
             state.pluginName = await input.showInputBox({
                 title: `${l10n.t('Create a new manifest')} - ${l10n.t('manifest name')}`,
                 step: step++,
@@ -576,7 +544,7 @@ class MultiStepInput {
                 step = await step(this);
             } catch (err) {
                 if (err === InputFlowAction.back) {
-                    if (onNavBack) { 
+                    if (onNavBack) {
                         onNavBack();  //Currently, step -= 2 passed as onNavBack because of using postfix increment in steps in the input functions
                     }
                     this.steps.pop();
